+++ src/player/ust/player.rs
@@ -249,7 +249,7 @@ impl FormatPlayer for USTPlayer {
         }
 
         data.speed = 6;
-        data.tempo = module.tempo as usize;
+        data.tempo = module.tempo as f32;
         data.time  = 0.0;
 
         let pan = match self.options.option_int("pan") {
